Correct build error
Replace incorrect type value "unsigned int" for "size_t" of variable
length in function SocketFrontEnd::socket_set_imdata
(ism/modules/frontend/scim_socket_frontend.cpp).

Bug-Tizen: PTREL-731

+++ ism/modules/frontend/scim_socket_frontend.cpp
@@ -1548,7 +1548,7 @@ SocketFrontEnd::socket_set_imdata (int /*client_id*/)
 {
     uint32 siid;
     char *imdata = NULL;
-    unsigned int length;
+    size_t length;
 
     SCIM_DEBUG_FRONTEND (2) << __func__ << "\n";
